pyproject.toml
[tool.poetry]
name = "discordBot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
PyNaCl = "^1.5.0"
jikanpy = "^4.3.2"
python-dotenv = "^0.20.0"
numpy = "^1.22.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
